+++ core/src/main/java/com/afs/dragbits/Camara/SeguimientoJugador.java
@@ -31,4 +31,5 @@ public class SeguimientoJugador {
     public void resize(float ancho, float alto) {
         camara.setToOrtho(false, ancho, alto);
     }
+
 }